fix: applique les retours de revue Copilot
- rend intitule_long nullable aussi dans le schéma v1
  /appointment-request/context/create (oublié lors du fix v2)
- factorise le garde-fou de reload (sessionStorage protégé par try/catch)
  entre ErrorComponent et instrumentation-client au lieu de dupliquer une
  version non protégée
- corrige le commentaire sur la portée réelle du reload automatique
  (ChunkLoadError et Server Action uniquement, pas les TypeError génériques)
